Implement debugging for ImGUI
The implementation of the debug renderer has been replaced

+++ src/xrGame/debug_renderer.h
@@ -14,7 +14,7 @@
 
 class CDebugRenderer {
 private:
-			void	add_lines		(Fvector const *vertices, u32 const &vertex_count, u16 const *pairs, u32 const &pair_count, u32 const &color);
+			void	add_lines		(Fvector const *vertices, u32 const &vertex_count, u32 const *pairs, u32 const &pair_count, u32 const &color);
 public:
 	IC		void	render			();
